MapContextEditor: Separate MapContextLayerTree component.
