+++ frontend/src/Components/MapContextEditor/MapContextLayerTree/MapContextLayerTree.css
@@ -1,0 +1,13 @@
+.mapcontext-layertree-node {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: row;
+  width: 100%;
+}
+
+.mapcontext-layertree-node-title,
+.mapcontext-layertree-node-actions {
+  display: flex;
+  flex-direction: row;
+  align-items: stretch;
+}
